Move automated transactions list to reg_admin
