+++ uber/templates/reg_admin/automated_transactions.html
